+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.2.18] - 2021-10-27
+### Added
+- Add optional sublabel to Confirmation component and center label on radio buttons
+- Add className prop to Message component and Button component for use with styled components
+- Make onClick prop on Button component optional for cases when using type="submit" prop in forms
+## [1.2.17] - 2021-10-21
+### Added
+- Visual updates to input components
+- Add optional onBlur prop to Confirmation component
+- Add optional type prop to Button component
+## [1.2.16] - 2021-10-15
+### Added
+- Add missing optional flag value prop in Dropdown component
+## [1.2.15] - 2021-10-15
+### Added
+- Add missing event handling props to Dropdown component
+
+## [1.2.14] - 2021-10-14
+### Added
+- Add error props and event handling props to input components
+
 ## [1.2.13] - 2021-10-12
 ### Added
 - Export new confirmation radio buttons component
@@ -151,6 +172,11 @@
 
 ### Changed
 - Updated gap and styles on Row component
+[1.2.18]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.17...v1.2.18
+[1.2.17]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.16...v1.2.17
+[1.2.16]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.15...v1.2.16
+[1.2.15]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.14...v1.2.15
+[1.2.14]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.13...v1.2.14
 [1.2.13]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.12...v1.2.13
 [1.2.12]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.11...v1.2.12
 [1.2.10]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.9...v1.2.10